telemetry-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
pymysql==1.1.0
pydantic==2.6.4
pydantic-settings==2.2.1
paho-mqtt==1.6.1
influxdb-client==1.41.0
redis==5.0.3
requests==2.31.0
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
